Rename blacklist to switchable_graphics
