+++ crates/hardware/src/switchable_graphics.rs
